confidence type
change from my assumed int to float32, which is what it appears to be

+++ core/fastsync.go
@@ -39,7 +39,7 @@ type FSArticleEnrichment struct {
 
 type FSDocType struct{
 	Type string `json:"type"`
-	Confidence int `json:"confidence"`
+	Confidence float32 `json:"confidence"`
 }
 
 type FSLanguage struct {
